refactor(ui): разделён пайплайн рендеринга reader-контента
ArticleScreenContentRenderer оставлен thin orchestration layer для выбора body source и fallback state
body block models, payload renderer, HTML block renderer, inline text renderer, media URL resolver и HTML utilities вынесены в отдельные файлы
поведение reader rendering сохранено, но границы parsing, media resolution и final body assembly стали явнее
